chore: New Dockerfile to build arm v6 packages natively
Implemented a Dockerfile emulating arm v6 bootstrapped with raspbian to
build debian packages natively.
It is built by calling the build.sh script, which creates an image that
boostraps raspbian using debootstrap. The resuling rootfs is extracted from
the image, and then used as the filesystem for the final image that
actually emulates armhf.

Ticket: MEN-6875

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>