chore: Upgrade Raspberry Pi OS tester image for Bullseye release
In the context of deprecating Debian Buster, we need to bump the tester
image to the oldest still supported release.

At the time of Bullseye release, the upstream project renamed the OS
from "Raspbian" to "Raspberry Pi OS" (`raspios` for short) so let's stick
to that for our naming of the image and the internal variables. For the
Docker image name, remove the "latest" part as it is not meaningful.

The rest of the changes [I hope that] are self-explanatory.

Ticket: QA-691

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>