ci: Switch client acceptance tests image to Python based
Which is a Debian based image with Python installed from source in
`/usr/local` (which does not conflict with system's Python).

This aims to address two issues:

* Latest Ubuntu and Debian images follow PEP 668 marking the Python
  environment as "externally managed" and preventing pip to install
  modules and apps normally. It can be ignored (not recommended) or
  workarrounded with virtual environments (the proper way, although
  cumbersome for containers)

* Ubuntu is moving more and more towards snaps, so we will be forced to
  mix deb packages and snaps sooner or later, which doesn't make much
  sense for containers.

As a bonus, we get this image presumably updated on newer Debian
releases and through dependabot we will catch when our own packages
(namely mender-artifact) need to be updated for the new distro.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>