build: Add mender-dist-packages builder image for Debian armhf
Important: this commit modifies the name of the builder images.

The purpose is to better show what the image is meant to (simulated
build or cross compiling build) and, most importantly, not reuse the
old Debian armhf (what was really Raspberry Pi OS armhf) with the new
Debian armhf (the "good" one).

The actual changes are minimal: remove the special handling of the
Dockerfile for Debian armhf and add some extra checks in the wrapper to
make sure we don't build unexpected images.

Ticket: MEN-7731

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>