test: Increase timeout waiting for container to 15 minutes
We started to see random failures in mender-dist-packages acceptance
tests on this wait for boot. This test suite run arm containers on x86
through emulation.

While in the past 3 minutes showed to be enough, it seems like some
either in the image or in the infra made the tests run into this limit.

Ticket: QA-929

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>