test: Modify Mender configured fixture to install 5.0.2 or newer
The previous implementation was broken for `mender-client4`.

On the one hand, any client between 4.0.0 and 5.0.1 don't exist in the
new device-components repository.

On the other hand, default `apt` installs recommended packages
(`mender-update recommends `mender-flash` and `mender-setup`).

Independently of the above, still supporting to pre-install the client
3.5 has little value nowadays, so let's stick to a recent client. At the
time of writing this the only LTS supported client is 5.x.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>